% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvdr-package.R
\docType{package}
\name{qvdr-package}
\alias{qvdr}
\alias{qvdr-package}
\title{qvdr package}
\description{
Provides an R interface to the qvdreader C++ library for reading Qlik QVD files, exposing both metadata and table data.
}
\author{
\strong{Maintainer}: Michael Developer \email{michael@example.com}

Other contributors:
\itemize{
  \item Devin Smith (Author of qvdreader library) [copyright holder]
}

}
\keyword{internal}
